# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vectric Blog Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List uploaded media files
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Media files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaList"

    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUpload"
      responses:
        "201":
          description: Uploaded media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # --- MESSAGES ---
...
components:
  schemas:
...
    MediaUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    Media:
      type: object
      required: [id, filename, url, size, createdAt]
      properties:
        id: { type: integer }
        filename: { type: string }
        originalName: { type: string }
        url: { type: string }
        mimeType: { type: ["string", "null"] }
        size: { type: integer }
        width: { type: ["integer", "null"] }
        height: { type: ["integer", "null"] }
        createdAt: { type: string }
    MediaList:
      type: object
      required: [media, total, page, limit]
      properties:
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # MESSAGES
...